% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq_data_raw}
\alias{eq_data_raw}
\title{NOAA Significant Earthquake Data (entire DB up until April 2017)}
\format{
A dataframe with columns described here:
 \url{https://www.ngdc.noaa.gov/nndc/struts/results?&t=101650&s=225&d=225}
}
\source{
U.S. National Oceanographic and Atmospheric Administration (NOAA)
 \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
eq_data_raw
}
\description{
NOAA Significant Earthquake Data (entire DB up until April 2017)
}
\examples{
\dontrun{
eq_data_raw
}
}
\keyword{datasets}
