% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean earthquake location names}
\usage{
eq_location_clean(x)
}
\arguments{
\item{x}{A character vector containing earthquake
location names.}
}
\value{
A character vector containing cleaned
location names.
}
\description{
Removes the country name and colon from the NOAA
LOCATION_NAME variable and converts the remaining
location name to title case.
}
\examples{
eq_location_clean("JORDAN: BAB-A-DARAA,AL-KARAK")

}
