% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Map earthquake epicenters}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A data frame containing cleaned earthquake data.}

\item{annot_col}{Character string giving the name of the column
to use for popup annotations.}
}
\value{
A leaflet map object.
}
\description{
Creates an interactive leaflet map showing earthquake epicenters.
Each earthquake is represented by a circle whose radius is
proportional to the earthquake magnitude.
}
\details{
The column used for the popup annotation is selected using
the \code{annot_col} argument.
}
\examples{
\dontrun{
mexico_data <- readr::read_delim(
  "earthquakes.tsv.gz",
  delim = "\t"
) |>
  eq_clean_data() |>
  dplyr::filter(
    COUNTRY == "MEXICO",
    !is.na(DATE)
  )

eq_map(
  mexico_data,
  annot_col = "DATE"
)
}

}
