% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{get_timeline}
\alias{get_timeline}
\title{get_timeline}
\usage{
get_timeline(data_raw, clist = "ALL", xmin, xmax)
}
\arguments{
\item{data_raw}{A data table containing NOAA Earthquake data}

\item{clist}{Character/Vector of grouping names eg "USA" from COUNTRY column.}

\item{xmin}{POSIXct date - minimum date for timeline}

\item{xmax}{POSIXct date - maximum date for timeline}
}
\value{
A ggplot2 graphical object displaying timeline of earthquakes data
}
\description{
A wrapper function to help generate timeline visualisations easier.
}
\examples{
\dontrun{
get_timeline(eq_data_raw, c("USA","IRAN"),"1970-01-01","2016-01-01")
}
}
