% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{get_timeline_label}
\alias{get_timeline_label}
\title{get_timeline_label}
\usage{
get_timeline_label(data_raw, clist = "ALL", xmin, xmax, n_max = 5)
}
\arguments{
\item{data_raw}{A data table containing NOAA Earthquake data}

\item{clist}{Character/Vector of grouping names eg "USA" from COUNTRY column.}

\item{xmin}{POSIXct date - minimum date for timeline}

\item{xmax}{POSIXct date - maximum date for timeline}

\item{n_max}{Integer value to control number of labels per group to show}
}
\value{
A ggplot2 graphical object displying timeline earthquate data with labels.
}
\description{
A wrapper function to help generate timeline (with labels) visualisations easier.
}
\examples{
\dontrun{
get_timeline_label(eq_data_raw, c("USA","CHINA"),"2010-01-01","2016-01-01", n_max = 5)
}
}
