% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA significant earthquake data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A data frame containing NOAA significant
earthquake data.}
}
\value{
A data frame containing the cleaned earthquake data.
}
\description{
Creates a DATE variable from YEAR, MONTH, and DAY,
converts latitude and longitude to numeric values,
and cleans earthquake location names.
}
