% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(eq_data)
}
\arguments{
\item{eq_data}{A data table containing NOAA Earthquake data}
}
\value{
A date frame with a new column called date (in POSIXct format) and longitude and latitude
columns formated as numeric.
}
\description{
Takes the raw data set and adds new columns "date", "longitude" and "latitude".
}
\examples{
library(dplyr)
USA <- eq_clean_data(eq_data_raw) \%>\% dplyr::filter(COUNTRY \%in\% "USA")
USA_IRAN <- eq_clean_data(eq_data_raw) \%>\%
  dplyr::filter(COUNTRY \%in\% c("USA","IRAN"))

}
