% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(eq_data)
}
\arguments{
\item{eq_data}{A data table containing NOAA Earthquake data}
}
\value{
An interactive map displaying earthquate location for a given country with custom popup.
}
\description{
A function to generate a custom popup box for a selected earthquake showing location,
magnitude and total deaths.
}
\examples{
library(dplyr)
eq_clean_data(eq_data_raw) \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\% eq_map(annot_col="popup_text")

}
