% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(eq_data, annot_col)
}
\arguments{
\item{eq_data}{A data table containing NOAA Earthquake data}

\item{annot_col}{A column found in \code{eq_data} to annotate earthquake marker}
}
\value{
An interactive map displaying earthquate location for a given country with user defined popup.
}
\description{
A function to generate an interactive map showing earthquakes for a particular country.
The user specifies a column from the data which the earthquake is to be annotated by eg date.
}
\examples{
library(dplyr)
library(lubridate)
eq_clean_data(eq_data_raw) \%>\%
   dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) \%>\% eq_map(annot_col="date")
}
