% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\value{
ggplot2 graphical object
}
\description{
A ggplot2 graphical function to plot a timeline of earthquakes from cleaned data.
The plot indicates the magnitude of each earthquake and number of deaths.
}
\section{Aesthetics}{

\code{geom_timeline} understands the following aesthetics:
\itemize{
  \item \code{x} date
  \item \code{y} latitude
  \item \code{xmin} minimum date for earthquakes
  \item \code{xmax} maximum date for earthquakes
  \item \code{size} used to size shape based on magnitude of earthquake eg EQ_PRIMARY
  \item \code{fill} used to colour shape based on number of deaths eg DEATHS
  \item \code{colour} used to colour shape based on number of deaths eg DEATHS
}
}

\examples{
library(dplyr)
library(ggplot2)
library(lubridate)
eq_clean_data(eq_data_raw) \%>\% eq_location_clean() \%>\%
dplyr::filter(COUNTRY \%in\% c("USA","IRAN"))  \%>\%
    ggplot2::ggplot() +
    geom_timeline(aes(x = date,
                      y = COUNTRY,
                      colour = DEATHS,
                      size = EQ_PRIMARY,
                      fill = DEATHS,
                      xmin = lubridate::ymd_hm("2000-01-01",truncated = 2),
                      xmax = lubridate::ymd_hm("2016-01-01",truncated = 2)))

}
