% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\value{
A ggplot2 graphical object for labelling plots generated with geom_timeline.
}
\description{
A ggplot2 graphical function that adds labels to earthquakes visualised.
There is an option to select the "n" largest earthquakes by magnitude to which to apply the labels.
Best used with `eq_location_clean`.
}
\section{Aesthetics}{

\code{geom_timeline_label} understands the following aesthetics:
\itemize{
  \item \code{x} date
  \item \code{y} (optional) aes can be used to group output eg by COUNTRY
  \item \code{location} aes used to selection labels eg LOCATION_NAME
  \item \code{xmin} minimum date for earthquakes
  \item \code{xmax} maximum date for earthquakes
  \item \code{size} aes used to indicate size eg EQ_PRIMARY
  \item \code{n_max} the top n number of labels to show based on size aes, defaults to n = 5
}
}

\examples{
library(dplyr)
library(ggplot2)
library(lubridate)
eq_clean_data(eq_data_raw) \%>\% eq_location_clean() \%>\%
dplyr::filter(COUNTRY \%in\% c("USA","IRAN"))  \%>\%
    ggplot2::ggplot() +
    geom_timeline(aes(x = date,
                      y = COUNTRY,
                      colour = DEATHS,
                      size = EQ_PRIMARY,
                      fill = DEATHS,
                      xmin = lubridate::ymd_hm("2000-01-01",truncated=2),
                      xmax = lubridate::ymd_hm("2016-01-01",truncated=2))) +
    geom_timeline_label(aes(x = date,
                            location = LOCATION_NAME,
                            xmin = lubridate::ymd_hm("2000-01-01",truncated=2),
                            xmax = lubridate::ymd_hm("2016-01-01",truncated=2),
                            size=EQ_PRIMARY,n_max=5,y=COUNTRY))

}
